cambios en fk

+++ Kinematics/src/InverseK.java
@@ -10,23 +10,22 @@ public class InverseK {
     }
 
 
-    public double[] getAngles(float coord[]) {
+    public double[] getAngles(double coord[]) {
 
         double Q[] = new double[3];
 
 
         System.out.println("\nX " + coord[0]+ " Y "+ coord[1] + " Z "+ coord[2]);
 
-        float r =(float) Math.sqrt(Math.pow(coord[0],2)+Math.pow(coord[1],2));
-       // float D = (float) Math.sqrt(Math.pow(coord[3]-Long_arm[0],2) + Math.pow(r,2));
+        double r = Math.sqrt(Math.pow(coord[0],2)+Math.pow(coord[1],2));
 
 
 
-        float gamma = (float) atan2(coord[1], coord[0]);
-        float beta = (float) acos((Math.pow(coord[2]-Long_arm[0],2) + Math.pow(coord[0],2) +  Math.pow(coord[1],2) - Math.pow(Long_arm[2],2) - Math.pow(Long_arm[1],2) )
+        double gamma =  atan2(coord[1], coord[0]);
+        double beta =  acos((Math.pow(coord[2]-Long_arm[0],2) + Math.pow(coord[0],2) +  Math.pow(coord[1],2) - Math.pow(Long_arm[2],2) - Math.pow(Long_arm[1],2) )
                 / 2*Long_arm[1]*Long_arm[2] );
 
-        float alpha = (float) (atan2(r,coord[2]-Long_arm[0]) - atan2(Long_arm[1] + Long_arm[2]* cos(beta),Long_arm[2]*sin(beta)));
+        double alpha =  (atan2(r,coord[2]-Long_arm[0]) - atan2(Long_arm[1] + Long_arm[2]* cos(beta),Long_arm[2]*sin(beta)));
 
 
         Q[0] = gamma;
